requirements.txt

streamlit>=1.36
mysql-connector-python>=9.0.0
bcrypt>=4.1.3
